odsc/requirements-dev.txt
pytest
pytest-flask
coverage
tox
